Modify README for empty taxonomy WAR
For empty taxonomy a user can achieve that by having an empty
(truly empty, not just "empty" entries) taxonomy section. The
README is a bit misleading by stating the bug here cannot be
worked around. While its description is technically correct,
augment the section to specify that an empty taxonomy is indeed
possible.